MotionManifolds.jl

Overview

Documentation for MotionManifolds.jl, a small library to manipulate matrix Lie groups frequently encountered in robotics.

A working minimal example:

julia> skew(SA_F64[1,2,3])
3×3 SMatrix{3, 3, Float64, 9} with indices SOneTo(3)×SOneTo(3):
  0.0  -3.0   2.0
  3.0   0.0  -1.0
 -2.0   1.0   0.0

Contents:

MotionManifolds.jl

Overview

Library

Manifold manipulations

Computation of derivatives

Utils

Library

The convention is that Lie groups are denoted with upper case letters, while their respective Lie algebra are denoted with lower case letters.

Special Orthogonal 2:

MotionManifolds.SO2 — Type

SO2

MotionManifolds.so2 — Type

so2 { w::Float64 }

Special Euclidian 2

MotionManifolds.SE2 — Type

SE2 { t::SVector{2,Float64}, rot::SO2 }

Possible constructors: SE2(x,y,th), SE2(t, rot), SE2(t, th), SE2(X::SMatrix)

MotionManifolds.se2 — Type

se2 {  vx::Float64, vy::Float64, w::Float64  }

Special Orthogonal 3

Special Orthogonal 3:

MotionManifolds.SO3 — Type

struct SO3
  u::SVector{3,Float64}
  w::Float64
  R::SMatrix{3,3,Float64,9}
end

Possible constructors: SO3(R); SO3(u,w); SO3(uw)

MotionManifolds.so3 — Type

struct so3
  u::SVector{3,Float64}
  w::Float64
end

$\mathfrak{so}(3)$

MotionManifolds.so3 — Method

so3(uw::SVector{3,Float64})

Constructor for $\mathfrak{so}(3)$

Special Euclidian 3

MotionManifolds.SE3 — Type

SE3 {  t::SVector{3,Float64}, rot::SO3  }

Possible constructors: SE3(t, rot),

MotionManifolds.se3 — Type

se3 { v::SVector{3,Float64}, w::so3 }

Possible constructors: se3(v,uw::SVector3), se3(v, uw::so3), se3(v, u, w)

Manifold manipulations

Base.:* — Function

*(rot1::SO2, rot2::SO2)

*(rot::SO2, t::SVector{2,Float64})

*(pose::SE2, t::SVector{2,Float64})

*(X1::SE2, X2::SE2)

*(Xr1::SO3,Xr2::SO3)

*(X1::SE3, X2::SE3)

*(rot::SO3, t::SVector{3,Float64})

*(rot1::SO2, rot2::SO2)

Base.inv — Function

inv(rot::SO2)

inv(X::SE2)

inv(rot::SO3)

inv(rot::SE3)

MotionManifolds.hat — Function

hat(uw::SVector{3,Float64}, ::Type{so3})::so3

hat(x::SVector{6,Float64}) -> se3

hat(w::Float64) -> so2

hat(tau::SVector{3,Float64})::se2

MotionManifolds.wedge — Function

wedge

Alias for hat.

MotionManifolds.vee — Function

vee(xr::so3)

vee(x::se3)

vee(w::so2)

vee(sk::se2)

MotionManifolds.exp_lie — Function

exp_lie(w::so2)

exp_lie(sk::se2) -> (; SE2, K1, w_sq)

MotionManifolds.log_lie — Function

log_lie(rot::SO2) -> so2

log_lie(X::SE2) -> se2

MotionManifolds.Exp — Function

Exp(wu::SVector{3,Float64}, ::Union{Type{SO3},Type{so3}})::SO3

$\mathfrak{so}(3)^\vee~\to~\mathrm{SO}(3)$

Exp(u::SVector{3,Float64}, w::Float64)::SO3

so3^\vee -> SO3

Exp(tau::SVector{6,Float64})::SE3

Exp(w::Float64)::SO2

so2^\vee -> SO2

Exp(tau::SVector{3,Float64})::SE2

se2^\vee -> SE2

MotionManifolds.Log — Function

Log(X::SO3) -> SVector3

Log(X::SE3) -> SVector6

Log(rot::SO2) -> SVector3

Log(X::SE2) -> SVector3

MotionManifolds.Adjm — Function

Adjm(Xr::SO3)

Adjm(X::SE3)

Adjm(rot::SO2)

The adjoint matrix.

Adjm(X::SE2)

Computation of derivatives

MotionManifolds.Jr — Function

Jr(xr::so3)

Jr(vw::se3)

Jr

MotionManifolds.Jrinv — Function

Jrinv(xr::so3)

Jrinv(vw::se3)

Jrinv

MotionManifolds.Jl — Function

Jl(xr::so3)

Jl(vw::se3)

Jl

MotionManifolds.Jlinv — Function

Jlinv(xr::so3)

Jlinv(vw::se3)

Jlinv

MotionManifolds.ExpAndJr — Function

ExpAndJr(tau::SVector{3,Float64}) -> (SE2, Jr_exp)

Returns the group manifold object directly from a vector representation of a Lie algebra element. Also returns the right Jacobian, as it allows to save some computations compared to doing the 2 things separately.

Utils

MotionManifolds.skew — Function

skew(a::SVector{3,Float64})

Compute the 3x3 skew matrix of a 3d vector

Example

julia> skew()
2×2 SMatrix{2, 2, Float64, 4} with indices SOneTo(2)×SOneTo(2):
 0.0  -1.0
 1.0   0.0

skew(a::Number=1)

Compute the 2x2 skew of a number (1 by default)

MotionManifolds.is_skew — Function

is_skew(M)

Check if a matrix is skewed.

MotionManifolds.to_matrix — Function

to_matrix(xr::so3)

to_matrix(Xr::SO3)

to_matrix(X::SE3)

to_matrix(r::SO2)

to_matrix(X::SE2)

to_matrix(w::so2)

to_matrix(sk::se2)

MotionManifolds.ecpi — Function

ecpi

Angle value of SO2 object, given in bounds [-π,π]